Fixed up a few comments/visibilities and moved uninitialized plugins to CLI crate.
